// src/app/api/contribuyente/models/relacionado-buscar-response.ts
/* tslint:disable */
/* eslint-disable */
export interface RelacionadoBuscarResponse {
  apeMaterno?: string;
  apePaterno?: string;
  contribuyenteNumero?: number;
  docIdentidadId?: number;
  estadoDsc?: string;
  index?: number;
  municipalidadId?: number;
  nombreCompleto?: string;
  nombres?: string;
  numDocIdentidad?: string;
  relContribuyenteNumero?: number;
  tipoDocumentoDsc?: string;
  tipoRelacionadoDsc?: string;
}